Fix Durable Object initializePromise example
In the previous version of the code, if the initialize() method throws
an exception synchronously (e.g.  because the `get()` isn't allowed due
to an ongoing `deleteAll()` operation), it could cause
`initializePromise` to get stuck with an exception in it due to the
order of assignments -- the assignment of `undefined` to
`initializePromise` would actually happen before the assignment of the
result of `initialize()` to `initializePromise`.

This formulation avoids that problem -- according to the relevant MDN
docs on Promise.then and Promise.catch, they're guaranteed to run
asynchronously.