[Teams] Add note about duplicate users for short term certificates (#2503)
In my opinion, important addition to this section since the restriction of usernames and email addresses being the same is of an issue than it might seem like (and its not immediately obvious).

I've been told doing this may lead to unintentional consequences with specific applications, but have had no trouble with this setup myself. This might be too sketchy and error prone to recommend in a security context though.