Remove the reference to `worktop/cookie` in the cookie parsing example. (#2989)
* Remove the reference to `worktop/cookie` in the cookie parsing example.

This doc gives a choice between 2 packages but isn't clear why one should be picked over the other (in fact, it claims they're the same). This confused a friend of mine who was looking at it, and I agreed it's confusing. This PR removes the reference to `worktop/cookie` in the example.

* Update products/workers/src/content/examples/extract-cookie-value.md

Co-authored-by: Kate Tungusova <70746074+deadlypants1973@users.noreply.github.com>

Co-authored-by: Kristian Freeman <kristian@kristianfreeman.com>
Co-authored-by: Kate Tungusova <70746074+deadlypants1973@users.noreply.github.com>