Clarify a bit of awkwardness in the KV limits documentation
We were saying that the free tier is limited to 1 GB of
"keys/namespace", which is kind of an odd framing. The limit is 1 GB of
data (whether stored in keys, or values, or metadata), and there's no
specific limit on the number of keys, just the 1 GB total storage
restriction.